make cuttinglabel's label equivalent to atom's label attr
previously cuttinglabel's label serves as the name/type of
cuttinglable: e.g., R, L. It becomes confusing since Atom has
label attr as well but for different purpose and we use lots
of methods originally designed for atom for cuttinglabel. so it
is better to make cuttinglabel's label attr equivalent to atom's
label attr, and create a new attr "name" that stores the name/type
of cuttinglabel.